Refactor to modules
