+++ src/main/java/com/github/hcsp/News.java
@@ -1,14 +1,10 @@
 package com.github.hcsp;
 
-import java.util.Date;
-
 public class News {
     private Integer id;
     private String url;
     private String content;
     private String title;
-    private Date createdAt;
-    private Date modifiedAt;
 
     public Integer getId() {
         return id;
@@ -40,21 +36,5 @@ public class News {
 
     public void setTitle(String title) {
         this.title = title;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public Date getModifiedAt() {
-        return modifiedAt;
-    }
-
-    public void setModifiedAt(Date modifiedAt) {
-        this.modifiedAt = modifiedAt;
     }
 }
